[functional] Created search_context package
