+++ test/functional/android/search_context/find_by_image_tests.py
@@ -22,8 +22,7 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
 from appium import webdriver
-
-from .helper import desired_capabilities
+from test.functional.android.helper import desired_capabilities
 
 
 @pytest.mark.skip(reason="Need to fix broken test")
