+++ test/functional/ios/search_context/find_by_ios_class_chain_tests.py
@@ -15,12 +15,10 @@
 import unittest
 
 from appium import webdriver
+from test.functional.ios.helper import desired_capabilities
 
-from .helper import desired_capabilities
 
-
-class FindByElementWebelementTests(unittest.TestCase):
-
+class FindByIOClassChainTests(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         desired_caps = desired_capabilities.get_desired_capabilities('UICatalog.app.zip')
@@ -31,19 +29,16 @@ class FindByElementWebelementTests(unittest.TestCase):
         self.driver.quit()
 
     def test_find_element_by_path(self):
-        el = self.driver.find_element_by_ios_predicate('wdName == "UICatalog"')
-        self.assertEqual('UICatalog', el.get_attribute('name'))
+        el = self.driver.find_element_by_ios_class_chain('XCUIElementTypeWindow/**/XCUIElementTypeStaticText')
+        self.assertEqual('Action Sheets', el.get_attribute('name'))
 
-        c_el = el.find_elements_by_ios_predicate('label == "Action Sheets"')
-        self.assertEqual('Action Sheets', c_el[0].get_attribute('name'))
-
-        c_el = el.find_elements_by_ios_class_chain('**/XCUIElementTypeStaticText')
-        self.assertEqual('Action Sheets', c_el[0].get_attribute('name'))
-
-        c_el = el.find_elements_by_accessibility_id('UICatalog')
-        self.assertEqual('UICatalog', c_el[0].get_attribute('name'))
+    def test_find_multiple_elements_by_path(self):
+        el = self.driver.find_elements_by_ios_class_chain('XCUIElementTypeWindow/*/*/*')
+        self.assertEqual(6, len(el))
+        self.assertEqual('UICatalog', el[0].get_attribute('name'))
+        self.assertEqual(None, el[1].get_attribute('name'))
 
 
 if __name__ == '__main__':
-    suite = unittest.TestLoader().loadTestsFromTestCase(FindByElementWebelementTests)
+    suite = unittest.TestLoader().loadTestsFromTestCase(FindByIOClassChainTests)
     unittest.TextTestRunner(verbosity=2).run(suite)
