+++ test/functional/ios/search_context/find_by_ios_predicate_tests.py
@@ -15,8 +15,7 @@
 import unittest
 
 from appium import webdriver
-
-from .helper import desired_capabilities
+from test.functional.ios.helper import desired_capabilities
 
 
 class FindByIOSPredicateTests(unittest.TestCase):
